registry_be/requirements.txt
asgiref
Django
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
cffi
cryptography
pycparser

